integration of cash accounts

+++ src/sample/CashAccount.java
@@ -14,13 +14,13 @@ public class CashAccount implements  Serializable{
 
     }
 
-    public void addFunds(int amount){
+    public void addFunds(double amount){
 
         balance += amount;
 
     }
 
-    public void subtractFunds(int amount){
+    public void subtractFunds(double amount){
 
         balance -= amount;
